Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.1
fpdf2==2.7.9
Werkzeug==2.3.7